refactor(export): split step_text.rs tests into step_text_tests.rs
2715d4bf8's quote-aware header scan pushed step_text.rs to 452 lines,
over the module_size_ratchet's 400-line LIMIT. Production code is only
~193 lines; the #[cfg(test)] block was the rest. Move it to a sibling
step_text_tests.rs (ratchet-exempt via the _tests.rs suffix), matching
the existing pattern in this crate (model_tests.rs, gltf_tests.rs,
frame_tests.rs, geom_tests.rs, model_inherit_tests.rs).

Pure code move: production code is byte-identical, and every test
still runs under its original path (step_text::tests::*).
